Refactor to use unit tests
